pyproject.toml
[project]
name = "fangclass-question-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pyzbar>=0.1.9",
    "qrcode-terminal>=0.8",
    "requests==2.32.3",
    "selenium>=4.39.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
